Remove linking liblldb to libpython
Co-authored-by: Uwe L. Korn <uwelk@xhochy.com>

liblldb doubles as an extension module too, but python
executable is linked to static python. When the extension module
is loaded both the shared libpython and the static libpython symbols
are loaded which leads to a crash in OSX.

To fix this, dependence on PYTHON_LIBRARY is removed from liblldb
and added to its interfaces to make sure the executables like
lldb works fine.

Thanks to feedback from @saulshanabrook