[runtimes] Probe for -nostdlib++ and -nostdinc++ with the C compiler
If they are detected, we add them to CMAKE_REQUIRED_FLAGS. All flags
in that variable are used both when testing compilation of C and C++
(and for detecting libraries, which uses the C compiler driver).

Therefore, to be sure we safely can add the flags to
CMAKE_REQUIRED_FLAGS, test for the option with the C language.

This should fix compilation with GCC; newer versions of GCC do
support the -nostdlib++ option, but it's only supported by the
C++ compiler driver, not the C driver. (However, many builds of
GCC also do accept the option with the C driver, if GCC was
compiled with Ada support enabled, see [1]. That's why this
issue isn't noticed in all configurations with GCC.)

[1] https://github.com/llvm/llvm-project/issues/90332#issuecomment-2325099254